Make a roses are red poem

Make a "roses are red ..." poem about an R package.

Usage

roses(
  pkg,
  hint = "",
  emoji = TRUE,
  chat = chat_openai(model = "gpt-3.5-turbo"),
  error_call = current_env()
)

prompt(pkg, hint = "", emoji = TRUE)

Arguments

pkg

A package

hint

extra information to add to the prompt

emoji

Should the poem include emojis ?

chat

A ellmer::Chat object, e.g. ellmer::chat_openai(). The default uses the 'gpt-3.5-turbo' model from OpenAI.

error_call

The execution environment of a currently running function, e.g. caller_env(). The function will be mentioned in error messages as the source of the error. See the call argument of abort() for more information.

Value

A lovely poem for your package crush

Examples

prompt("dplyr")
#> Make a 4 lines subtle "roses are red ..." poem to express your love for the R package "dplyr". Include a bunch of emojis. 

if (FALSE) { # \dontrun{
  # this needs the OPENAI_API_KEY environment variable
  # to be set. Visit https://openai.com/api/
  roses("dplyr")
} # }